multiplied gravity value by 2
i multiplied -9.81 by 2 since the jump was a bit like a jump on the moon. not sure if this change will stick, since i changed the value in the inspector rather than in code. but anyway, the jump is a little more realistic. if you want, change it again by mutliplying -9.81 by 3 in the inspector if you feel the jump is still a bit weird.