adjusted terrain and added ocean
made the terrain look a little like an island, and added an ocean surrounding it. if anyone can make the terrain look much more like an island, it would be much appreciated!! only the shape and all is island-like, so would be nice to make it look more like one.